+++ GPTifier/include/cli.hpp
@@ -4,8 +4,9 @@
 
 namespace cli {
 
+void exit_on_failure();
+
 ParamsRun get_opts_run(int argc, char **argv);
-ParamsShort get_opts_short(int argc, char **argv);
 ParamsModels get_opts_models(int argc, char **argv);
 ParamsEmbedding get_opts_embed(int argc, char **argv);
 ParamsFineTune get_opts_create_fine_tuning_job(int argc, char **argv);
